HAL: better exponential backoff for SleepForSpinning()
Sleeping is just plenty horrible and terribly creeples performance when
there is contention for an atomic lock.
Atomic locks are only useful when we assume little to no-collision, otherwise
using regular kernel primitives (like a mutex) will clearly benefit from OS scheduling.

+++ Source/Runtime/Core/Private/Allocator/MallocBinned.cpp
@@ -1100,7 +1100,7 @@ static NODISCARD NO_INLINE void* LargeRealloc_(void* const oldp, size_t newSize,
 //----------------------------------------------------------------------------
 FBinnedThreadCache_::~FBinnedThreadCache_() {
     // let other thread release blocks belonging to this thread, if any
-    for (size_t backoff = 0;;) {
+    for (i32 backoff = 0;;) {
         ReleaseDanglingBlocks_(*this);
         FPlatformProcess::SleepForSpinning(backoff);
         if (nullptr == DanglingBlocks)
